AI - A/requirements.txt
# Ana AI Assistant Requirements

# Core dependencies
PyQt5>=5.15.11
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.12.2
matplotlib==3.7.2
numpy>=1.20.0
requests>=2.25.1

# Google Authentication
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# AI and ML
openai==1.3.0
tensorflow==2.12.0
opencv-python==4.8.0.76
scikit-learn==1.3.0
face-recognition==1.3.0

# Voice and audio
SpeechRecognition>=3.8.1
pyttsx3>=2.90
pyaudio>=0.2.11
elevenlabs==0.2.24
pydub==0.25.1
pvporcupine>=2.2.0  # Wake word detection

# Weather and External APIs
requests>=2.25.1  # For weather API calls

# GitHub integration
PyGithub==1.59.0

# Security and encryption
cryptography==41.0.3
pyjwt==2.8.0
bcrypt==4.0.1

# UI and Graphics
PyQt5-SVG>=5.15.0  # SVG rendering support
pillow==10.0.0  # Image processing

# Health integration
# Note: Samsung Health API requires developer account and API key

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp>=3.8.5  # Async HTTP requests
threading  # Built-in
asyncio  # Built-in
datetime  # Built-in
enum  # Built-in 